+++ src/parser/inblock.test.ts
@@ -21,20 +21,6 @@ describe('Unit: parser/inblock', () => {
       assert.deepStrictEqual(inspect(parser(' @a#1')), [[' ', '<a class="channel" rel="noopener">@a#1</a>'], '']);
     });
 
-    it('account', () => {
-      assert.deepStrictEqual(inspect(parser('@a')), [['<a class="account" rel="noopener">@a</a>'], '']);
-      assert.deepStrictEqual(inspect(parser('@a@')), [['@', 'a', '@'], '']);
-      assert.deepStrictEqual(inspect(parser('@a@b')), [['@', 'a', '@', 'b'], '']);
-      assert.deepStrictEqual(inspect(parser('@a,@b')), [['<a class="account" rel="noopener">@a</a>', ',', '<a class="account" rel="noopener">@b</a>'], '']);
-      assert.deepStrictEqual(inspect(parser('@@a')), [['@', '@', 'a'], '']);
-      assert.deepStrictEqual(inspect(parser('a@b')), [['a', '@', 'b'], '']);
-      assert.deepStrictEqual(inspect(parser('a@@b')), [['a', '@', '@', 'b'], '']);
-      assert.deepStrictEqual(inspect(parser('_@a')), [['_', '<a class="account" rel="noopener">@a</a>'], '']);
-      assert.deepStrictEqual(inspect(parser('*@a*')), [['<em>@a</em>'], '']);
-      assert.deepStrictEqual(inspect(parser('(@a)')), [['(@a)'], '']);
-      assert.deepStrictEqual(inspect(parser(' @a')), [[' ', '<a class="account" rel="noopener">@a</a>'], '']);
-    });
-
     it('hashtag', () => {
       assert.deepStrictEqual(inspect(parser('#a#')), [['#', 'a', '#'], '']);
       assert.deepStrictEqual(inspect(parser('#a#b')), [['#', 'a', '#', 'b'], '']);