+++ src/parser/inline.ts
@@ -49,3 +49,4 @@ export { label, isGroup, isFixed } from './inline/extension/label';
 export { link } from './inline/link';
 export { media } from './inline/media';
 export { uri } from './inline/autolink/uri';
+export { account } from './inline/autolink/account';