+++ src/parser/inline/autolink.ts
@@ -1,7 +1,9 @@
 ﻿import { AutolinkParser } from '../inline';
 import { union } from '../../combinator';
 import { uri } from './autolink/uri';
+import { account } from './autolink/account';
 
 export const autolink: AutolinkParser = union([
   uri,
+  account,
 ]);