+++ src/parser/inline/autolink/account.ts
@@ -1,4 +1,4 @@
-﻿import { AutolinkParser } from '../../inblock';
+﻿import { AutolinkParser } from '../../inline';
 import { union, some, subline, focus } from '../../../combinator';
 import { unescsource } from '../../source/unescapable';
 import { html } from 'typed-dom';
@@ -9,5 +9,5 @@ export const account: AutolinkParser.AccountParser = subline(union([
     some(unescsource)),
   focus(
     /^@[a-zA-Z0-9]+(?:-[0-9a-zA-Z]+)*(?!@)/,
-    source => [[html('a', { class: 'account', rel: 'noopener' }, source)], ''])
+    source => [[html('a', { class: 'account', rel: 'noopener' }, source)], '']),
 ]));